requirements.txt
astroid==1.6.3
Django==2.0.4
djangorestframework==3.8.2
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysqlclient==1.3.12
pkg-resources==0.0.0
psycopg2==2.7.4
pylint==1.8.4
pytz==2018.4
six==1.11.0
wrapt==1.10.11
